init commit 41

+++ search-models/src/main/java/ru/tecius/telemed/dto/response/SearchResponseDto.java
@@ -2,8 +2,8 @@ package ru.tecius.telemed.dto.response;
 
 import java.util.List;
 
-public record SearchResponseDto<E>(Integer totalElements,
-                                Integer totalPages,
+public record SearchResponseDto<E>(Long totalElements,
+                                   Long totalPages,
                                 Boolean moreRows,
                                 List<E> content) {
